Separated CSS & Assets fixtures
