+++ test/fixtures-assets.spec.js
@@ -69,92 +69,6 @@ beforeEach(done => {
 });
 
 
-it('CSS Modules', async() => {
-	const actualFiles = await testFixture({
-		rootFolder,
-		outputFolder,
-		mode: 'development',
-		entry: {
-			myapp: './css-modules/myapp.ts'
-		},
-		cssModules: true
-	});
-	const expectedFiles = [
-		'index.html',
-		'myapp.css',
-		'myapp.css.map',
-		'myapp.js',
-		'myapp.js.map'
-	];
-	expect(actualFiles.sort()).toEqual(expectedFiles.sort());
-
-	const browser = await puppeteer.launch();
-	try {
-		const page = await browser.newPage();
-		await page.goto(`http://localhost:${port}/`);
-		const found = await page.evaluate(() => {
-			/* global document */
-			/* global window */
-			const el = document.getElementById('hello');
-			if (el === null){
-				return '#hello not found';
-			}
-			const computed = window.getComputedStyle(el);
-			if (computed.getPropertyValue('color') !== 'rgb(0, 128, 0)'){
-				return 'Bad color';
-			}
-			return 'ok';
-		});
-		expect(found).toBe('ok', 'DOM tests');
-	} finally {
-		await browser.close();
-	}
-});
-
-
-it('CSS without CSS Modules', async() => {
-	const actualFiles = await testFixture({
-		rootFolder,
-		outputFolder,
-		mode: 'development',
-		entry: {
-			myapp: './css-no-modules/myapp.ts'
-		},
-		cssModules: false
-	});
-	const expectedFiles = [
-		'index.html',
-		'myapp.css',
-		'myapp.css.map',
-		'myapp.js',
-		'myapp.js.map'
-	];
-	expect(actualFiles.sort()).toEqual(expectedFiles.sort());
-
-	const browser = await puppeteer.launch();
-	try {
-		const page = await browser.newPage();
-		await page.goto(`http://localhost:${port}/`);
-		const found = await page.evaluate(() => {
-			/* global document */
-			/* global window */
-			const el = document.getElementById('hello');
-			if (el === null){
-				return '#hello not found';
-			}
-			const computed = window.getComputedStyle(el);
-			if (computed.getPropertyValue('color') !== 'rgb(0, 128, 0)'){
-				return 'Bad color';
-			}
-			return 'ok';
-		});
-		expect(found).toBe('ok', 'DOM tests');
-	} finally {
-		await browser.close();
-	}
-});
-
-
 it('Assets', async() => {
 	const actualFiles = await testFixture({
 		rootFolder,
@@ -302,51 +216,3 @@ it('Copy Patterns', async() => {
 	];
 	expect(actualFiles.sort()).toEqual(expectedFiles.sort());
 });
-
-
-it('CSS Reset', async() => {
-	const actualFiles = await testFixture({
-		rootFolder,
-		outputFolder,
-		mode: 'development',
-		entry: {
-			myapp: './css-reset/myapp.ts'
-		},
-		polyfills: [
-			'./css-reset/reset.css'
-		]
-	});
-	const expectedFiles = [
-		'index.html',
-		'myapp.css',
-		'myapp.css.map',
-		'myapp.js',
-		'myapp.js.map'
-	];
-	expect(actualFiles.sort()).toEqual(expectedFiles.sort());
-
-	const browser = await puppeteer.launch();
-	try {
-		const page = await browser.newPage();
-		await page.goto(`http://localhost:${port}/`);
-		const found = await page.evaluate(() => {
-			/* global document */
-			/* global window */
-			const el = document.createElement('div');
-			const computed1 = window.getComputedStyle(el);
-			if (computed1.getPropertyValue('color') === 'rgb(0, 0, 128)'){
-				return 'Unexpected color before appending';
-			}
-			document.body.appendChild(el);
-			const computed2 = window.getComputedStyle(el);
-			if (computed2.getPropertyValue('color') !== 'rgb(0, 0, 128)'){
-				return 'Bad color';
-			}
-			return 'ok';
-		});
-		expect(found).toBe('ok', 'DOM tests');
-	} finally {
-		await browser.close();
-	}
-});
-
